+++ model/src/commonMain/kotlin/dev/msfjarvis/claw/model/User.kt
@@ -10,7 +10,7 @@ class User(
   @SerialName("is_admin") val isAdmin: Boolean,
   val about: String,
   @SerialName("is_moderator") val isModerator: Boolean,
-  val karma: Long = 0,
+  val karma: Int = 0,
   @SerialName("avatar_url") val avatarUrl: String,
   @SerialName("invited_by_user") val invitedByUser: String,
   @SerialName("github_username") val githubUsername: String? = null,
